requirements.txt
django
django-configurations
django-allauth
requests
south
fabric